refactor: 헤더 및 메인 컴포넌트 폴더구조 수정
