+++ src/components/layout/main/main.module.scss
